% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.R
\docType{data}
\name{nswcps}
\alias{nswcps}
\title{National Supported Work - Current Population Survey (NSW-CPS)}
\format{
A data frame with 11 variables:
\describe{
\item{\code{treated}}{(1 if treated, 0 if not treated)}
\item{\code{age}}{age}
\item{\code{age2}}{age-squared}
\item{\code{educ}}{years of education}
\item{\code{black}}{(1 if black, 0 otherwise)}
\item{\code{hispanic}}{(1 if Hispanic, 0 otherwise)}
\item{\code{married}}{(1 if married, 0 otherwise)}
\item{\code{nodegree}}{(1 if no degree, 0 otherwise)}
\item{\code{re74}}{earnings in 1974}
\item{\code{re75}}{earnings in 1975}
\item{\code{re78}}{earnings in 1978}
}
}
\usage{
nswcps
}
\description{
Data are draws from Robert Lalonde
"Evaluating the Econometric Evaluations of Training Programs," American Economic Review, Vol. 76, pp. 604-620
}
\keyword{datasets}
